Readd text truncate on node header
