+++ src/assets/scss/_base.scss
@@ -86,3 +86,11 @@ $buffers-colors: (
   border-style: dashed !important;
   background-color: transparent !important;
 }
+
+.line-clamp-2 {
+  display: -webkit-box;
+  -webkit-line-clamp: 2;
+  -webkit-box-orient: vertical;
+  text-overflow: ellipsis;
+  overflow: hidden;
+}
